Debugging the create anime route! When experiencing the 405 EPIPE error, I have to remember to add the methods attribute specifying what type of method it is if not the default GET.
A mini but significant challenge I experience was the infinite recursion loop when posting a new anime that has the user who posted the anime.
The importance of to_dict_basic() is to prevent infinite nested loop between the anime having the user and the user having the anime they posted.