Have the checkbox return an array of arrays representing [anime.id, anime.title, boolean checkbox].
Having the checkbox be true means the character is associated with the anime.title and will have its id be associated to the anime.id in the anime_characters table once the form is submitted.

As of now, the function for associating the character id with the anime.id is NOT yet developed!