Successfully modify the addCharacterToAnime router to update the character's role if it already exist. I believe it's best to create a separate router and thunk for updating the character role instead though.
UpdateCharacterFormPage can now call for an update to the character's role based on the user's choice.
If I were to create a separate router/thunk for update, then I will need to change it soon.
I also need to add a call for deleting character association once I create the thunk for it.

I have already created a router for deleting the character's association from an anime.
Still need to create some thunks for it.