Isolate the "Welcome" component to be exported for both the AnimeCatalogPage and CharacterCatalogPage.
Rely on the entire browser to scroll through the catalog list.
Would save the trouble for the users to scroll all the way down to reach the horizontal scrollbar if one catalog end up being too long vertically.

Give a little overview for the AnimeShoKan Project in the README.
Will continue on writing more...